test: isolate PSR-4 fixtures
