+++ tests/Fixture/Doubles/StaticMethodFixture.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Greenlight\Tests\Fixture\Doubles;
+
+interface StaticMethodFixture
+{
+    public static function lookup(): string;
+}
