+++ tests/Fixture/Expect/Node.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Greenlight\Tests\Fixture\Expect;
+
+final class Node
+{
+    public ?Node $next = null;
+}
