+++ tests/Fixture/Expect/Signal.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Greenlight\Tests\Fixture\Expect;
+
+enum Signal
+{
+    case Green;
+}
