+++ tests/Fixture/Expect/Suit.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Greenlight\Tests\Fixture\Expect;
+
+enum Suit
+{
+    case Hearts;
+    case Spades;
+}
